Thu, Feb 24, 2022, 9:49 PM +00:00

+++ CS50x/week4/lab4/volume_DONE/volume.c
@@ -43,7 +43,7 @@ int main(int argc, char *argv[])
     // TODO: Read samples from input file and write updated data to output file
 
     float buffer;
-    uint8_t actual_buffer;
+    int actual_buffer;
 
     while (fread(&buffer, sizeof(uint16_t), 1, input))
     {
